main container 내부를 friend container와 chat container 로 분리
